CMPLRLLVM-25285: Change HLS Streams overload type-
A recent change to LLVM made it so intrinsics not considered 'overloaded'
could not have 'overloaded' like names. Apparently the normal Pointer
Type of llvm_any_ty is NOT enough to cause it to be overloaded.

SO, this change instead overloads based on pointer-to-type.
(AnyPointerType<llvm_any_ty>).  The only real outcome of this is that
the name of the overloaded-part changes from 'structFoos' to
'p0_structFoos', but the type of the function itself doesn't change.

There was some slight additional changes required to make sure that the
type of 'bool' stayed as i1 instead of of i8 (to differentiate between a
char/byte and a bool) in the case of 'writes'.

Change-Id: I739f8371f5524792a2073cf24af7ce4dbad6f9fd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/305116
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Manna, Soumi <soumi.manna@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>