[OptReport] Introduce mechanism to represent unique remark ID in LoopOptRemark.
Every remark emitted in a LoopOptReport is expected to have a
unique diagnostic ID which will be used while cataloging various
remark messages. Currently LoopOptReport framework ignores these
IDs and emits LoopOptRemark in below format -
!1 = !{!"intel.optreport.remark", !<Remark Msg>, <Remark Args>}

This patch adds new interfaces in LoopOptReport to explicitly
add a remark's unique ID to LoopOptRemark metadata. The current
interfaces are indicated as legacy code and all users of
LoopOptReport are expected to switch to the new interfaces. VPlan
vectorizer has been updated accordingly as an example. The new
format for LoopOptRemark is -
!1 = !{!"intel.optreport.remark", !<Remark ID>, !<Remark Msg>, <Remark Args>}

Change-Id: I0add2a6cb0f5095a4deae096ce5d15dd0ed36f8f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/305367
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>