CMPLRLLVM-25349: Improvements to inlining report and InlineReport71.ll fix.
This makes several changes to the inlining report to accommodate changes to
inlining in the new pass manager, provide for default printing of all current
inlinings in one report per compilation on both the old and new pass managers,
and to ensure that quality messages generated by one invocation of the inliner
are not overwritten by lower quality messages from a later invocation. (On that
last point, the key issue was the "always inliner" could overwrite a
non-inlining message like "Inlining is not profitable" with "Callee is not
always_inline". This is now special-cased in both the classic and metadata
inlining reports.

A key change is that we maintain a set of active inliners, and only dump the
inlining report when the last of these is destructed. This is sufficient to
generate only a single inlining report per compilation. The inlining report
mask setting to enable additional inlining reports is no longer needed now,
and so has been removed. The original LIT test InlineReport72.ll which tested
that setting is no longer needed. I replaced it with a test that the quality
non-inlining messages were not replaced by "Callee is not always_inline".

This simpler approach has the added advantage that many Intel customizations
needed in the previous approach can be reverted.

Change-Id: I3b430d2e96ad52da23257f0e978f7f2261dc8211
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/305336
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>