CMPLRLLVM-19406: Fix verifier error on 549.fotonik with -g
When generating the debug info for the multiversioning conditional in
tiling and inlining for multiversioning, the DebugLocs for the loads must
point to the subgrogram in which the conditional is generated. To ensure
this, I now generate the DebugLoc using the subprogram of the call around
which the multiversioning will occur. Previously, I used the DebugLocs
from the loads which were being copied. This is incorrect, because those
are loads of globals which could appear in any subprogram.

Change-Id: Ib153b0330891032ebb485f2afbec24bbf13d8b30
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/305571
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>