[NFCI][VPlan] Refactor of vector variant matching code
The refactoring work here transfers the responsibility of vector
variant matching from VPlanCallVecDecision to VectorVariant and
TTI. VectorVariant has been updated to allow matching between
a variant created for the call against other variants specified
as function attributes (e.g., vector-variants). TTI is used to
verify that the appropriate ISA is selected since the target
features are not available to VPlan.

Change-Id: Ic8fd6880fe6aa41c313056cdeb9fed944437f4d9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/305597
Reviewed-by: Kelarev, Ivan <ivan.kelarev@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>