[VPlan] Correct assertion check in LLVM IR vector code generation
The issue was encountered while working on the fix for CMPLRLLVM-23628.
We were incorrectly asserting for type size to always be a multiple of
8 during generation of gather/scatters. This is not true for types such
as <2 x i1>. The fix was originally part of the fix for CMPLRLLVM-23628
but as suggested in the review was separated out into a different patch
as the two issues are essentially independent.

Jira :: CMPLRLLVM-24685

Change-Id: I8bb073b088284de7ddd3e7f7231e20250269abfa
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/306094
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>