[NFC] Remove redundant call to getMainExecutable() (#3018)
The call to getMainExecutable() function can be eliminated because the
path returned by that call is used only for error reporting, so the tool
can use argv[0] for this purpose as is without any additional processing.

This is a cherry-pick of https://github.com/intel/llvm/pull/3018
which is expected to resolve problem reported in CMPLRLLVM-25520.

Change-Id: Ifdd555c7eabbbbf1319b8fec4daf0cb401df3281
Signed-off-by: Sergey Dmitriev <serguei.n.dmitriev@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/306526