CMPLRLLVM-25542: deploy llvm-lib for Windows.
llvm-lib is just a copy/symlink of llvm-ar, but the code behaves
differently depending on the executable name. llvm-lib is required
for building LTO static libraries, and llvm-ar cannot provide
the same functionality. llvm-ar is deployed as part of
SYCL toolchain, but now we are also adding llvm-lib for xmain.

Change-Id: I0ff803bbdf0a2cb2c341d39180ee1932f54a046e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/306711
Tested-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Dmitriev, Valery N <valery.n.dmitriev@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>