[NFC][Loopopt] Assert that canon expression convert is one of sext/zext/trunc
This is a followup change to address Pankaj's review comment in
https://git-amr-2.devtools.intel.com/gerrit/#/c/306500/ where he
suggested adding an assert to check that a canon expression convert
is not a bitcast. Changes have been added to assert that a canon
expression convert is one of sext/zext/trunc in canon expression
print and verify methods.

Change-Id: Ief4ba26066f607811e7e15c383f0eb93bad72ab3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/307009
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>