[DTrans] CMPLRLLVM-25425: Add support for DTrans analysis to analyze parameters to functions marked with 'callback' metadata
Parameters to a function marked with 'callback' metadata may be forwarded to a
target function that can be checked for DTrans safety, instead of marking the
types with 'System object' or 'Address taken'. This adds the analysis to check
the parameter against the type that is expected by the function the parameter
will be forwarded to.

This is part 2 of the changes for CMPLRLLVM-25425.

Change-Id: Id1eac1cd9571bfa13e32711df6e412368d986e73
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/306844
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>