cmplrllvm-25484:fix generate "CHAIN" modifer which should not.
The problem is clang add an optimization in emitCombinedEntry to
eliminate unnecessary base map info which will skip adding base to
CombinedInfo.  This change misses up size of VarChain.

To fix this before call to emitCombinedEntry, save the map
argument's size of CombinedInfo. After the call use it to check
if base info is added then generate info for VarChain as
needed.

Change-Id: I1b9a15b612ba4c8e267eed553368ab35617c3c4b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/307473
Tested-by: Yu, Jennifer <jennifer.yu@intel.com>
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Reviewed-by: Huang, Zao <zao.huang@intel.com>