[clang-offload-bundler] Initialize member names in unbundled archive
Member names in the unbundled archive were not properly initialized
which could lead to random problems for tools working with these
archives later. This patch fixes this problem.

There is no LIT test in this patch because it is impossible to create
stable test which shows the problem.

This change is expected to fix problem reported in CMPLRLLVM-25656.
It is similar to https://github.com/intel/llvm/pull/3049.

Change-Id: Ibd3e8c99a34ba05f460fe108635b906957afafed
Signed-off-by: Sergey Dmitriev <serguei.n.dmitriev@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/307187
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Bobrovsky, Konstantin S <konstantin.s.bobrovsky@intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>