[DTrans] CMPLRLLVM-25425: Update support for DTrans analysis of 'callback' functions
When DTrans analysis is checking the parameter types that will be forwarded via a
'callback' function, the target function may be indirect. In this case, treat the
parameter conservatively by setting the "Unhandled use" safety flag.

This fixes a problem with: https://git-amr-2.devtools.intel.com/gerrit/#/c/306844/
which would have tried to dereference a nullptr in that case.

Change-Id: I26d5bc93fa53e11eff2668d39e64baf698158a39
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/307507
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>