[SYCL] Specialize atomic fetch_add for floating point types
The new EXT/SPV_EXT_shader_atomic_float_add SPIR-V extension
allows us to further specialize atomic::fetch_add() for
floating point types. In device mode, we'll now be creating
an external call to a built-in-like __spirv_AtomicFAddEXT().
This is similar to what is done for other atomic binary
instructions, e.g. the integer specialization of fetch_add()
being mapped onto __spirv_AtomicIAdd().

Furthermore, atomic::fetch_sub() is also re-implemented
to use __spirv_AtomicFAddEXT(), the added operand being
a negation of the original one.

The new implementation can be exposed if a dedicated macro is
defined: SYCL_USE_NATIVE_FP_ATOMICS. Otherwise, a fallback
is used, where the atomic operation is done via spinlock emulation.
At the moment of committing this, only Intel GPUs support the
"native" implementation, which relies on a SPIR-V extension.

Tests for the feature have been finalized in
https://github.com/intel/llvm-test-suite/pull/104.

This is a cherry-pick of https://github.com/intel/llvm/pull/2765.

Change-Id: Ie2d0b6f33817d51b152f878fb53b92dd04e138b3
Signed-off-by: Artem Gindinson <artem.gindinson@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/307825
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>