[SYCL][PL][L0]Fixes a deadlock that can occur with batching enabled.
Cherry pick https://github.com/intel/llvm/pull/3081 to xmain.

This change causes the event wait list for any command being
enqueued to be checked to see if it was issued into a different
queue than the command that will be waiting on it. If it was,
then for batching to work properly, the currently open batch for
that queue is closed and submitted for execution.

Change-Id: I993809553daa3fe89eb27ae3844629a06fc683ba
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308368
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>