CMPLRLLVM-25685: Fixed a bug in HIR DSE where load and store in the same HLInst are not ordered correctly.
Memrefs of a group in DSE were being ordered purely based on top sort number.
Sorting of load and store with same top sort number was arbitrary which is incorrect.

Lexically, loads execute before the store of the same HLInst.

HIRLocalityAnalysis::populateEqualityGroups() now guarantees that the returned refs
are in strict lexical order and DSE relies on this property.

Change-Id: I2a8f12805238df0d51ff52268a54100e32d67a25
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308475
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>