[SYCL] Clean up shared events from queues upon reaching a threshold (#3068)
Previously the shared ownership of some events (USM and those without
dependencies) by queues was released only upon calls to
queue::wait/wait_and_throw(). If the application exclusively uses
event::wait() instead, those events will be kept alive until the
destruction of the queue.

The patch addresses this problem by introducing a threshold on shared
events associated with each queue. If that threshold is exceeded upon
adding a new shared event for the queue, its events are queried for
their execution status and the completed ones are released.

Change-Id: I4ee2f100a5f2eec821be5d321bcefe8f5660d1d7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308565
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: Semenov, Sergey <sergey.semenov@intel.com>
Reviewed-by: Wang, Jing W <jing.w.wang@intel.com>
Reviewed-by: Huang, Zao <zao.huang@intel.com>