[-g][simulator] Run DebugInfoPass before KernelAnalysis
If a callee is not inlined and its caller is kernel, DebugInfo pass
inserts get_global_id call to the callee. Both callee and the kernel is
patched in barrier pass, causing an additional arg is added to the
kernel.

This is fixed by move DebugInfoPass before KernelAnalysis pass.
KernelAnalysis will set no_barrier_path to false for caller kernel. The
kernel loop will be created in Barrier instead of LoopCreator pass.

Another possible solution is to extend LoopCreator pass to support
get_global_id calls in un-inlined callee functions.

Change-Id: Ic174b83890abbbd0d92ae1344894df9aed4208fd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308544
Reviewed-by: Wenju He <wenju.he@intel.com>
Tested-by: Wenju He <wenju.he@intel.com>