[SYCL][PL][L0]Fixes a deadlock that can occur with batching enabled.
Cherry pick https://github.com/intel/llvm/pull/3081 to xmain-rel.

This change causes the event wait list for any command being
enqueued to be checked to see if it was issued into a different
queue than the command that will be waiting on it. If it was,
then for batching to work properly, the currently open batch for
that queue is closed and submitted for execution.

Change-Id: I4ab00b840729a55d99dc5dbb3401287fd58bd341
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308577
Reviewed-by: Kevin B. Smith <kevin.b.smith@intel.com>
Tested-by: Kevin B. Smith <kevin.b.smith@intel.com>