CMPLRLLVM-25769: Update classic inlining report after partial inlining
The classic inlining report was not updated after community partial inlining.
The callsites that were part of the partially inlined code did not get
replicated into the caller. This caused a segfault in 523.xalancbmk, when
the inlining report tried to access the inline report representation of a
replicated call site and it did not exist.

(I also fixed an omission in InlineReport77.ll with this change set.)

Change-Id: Ief451ae26a1a26a2cb209e8a5cf7815b840985ea
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308609
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>