CMPLRLLVM-25822: Perform additional profitability checks for unrolling deeply nested multi-exit loop with liveouts.
Since unrolling deeply nested multi-exit loops can cause register
pressure to increase, we perform additional profitability check
for them as we do for unknown loops. Unrolling these loops is
resulting in degradations in multiple SPEC benchmarks and this
approach is trying to restrict the number of loops getting
unrolled. We can look for a cost better model later on.

Change-Id: I9da1e80bf64a2ad9e8a9e77e572b5fac6ba1e7c7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308684
Tested-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Reviewed-by: Zhao, Chuck <chuck.zhao@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>