Restore old code to generate alignment assumptions
The community [switched to using bundle-based alignment
assumptions](https://reviews.llvm.org/rG78de729) before support for
them is added to ValueTracking. This broke proper handling of
alignment-related and nontemporal pragmas in xmain. As a workaround, I
temporarily restore the old code in xmain. The old code can be removed
as soon as [support for assumption bundles in
ValueTracking](https://reviews.llvm.org/D82703) is committed to llvm
and pulled down to xmain.

Change-Id: I861b51dd8c24e8eae65d161563c65aefc1c08a33
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308723
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>