[DTrans] CMPLRLLVM-25425: Relax function address-taken restriction on DTransFieldModRef collection
Any function that was address taken would cause all fields that are accessed by
that function or functions reachable from it to be disqualified from the
analysis. In cases where the address taken use is to pass a function pointer to
be invoked by another function the analysis will be possible by treating the
address taken function as a callee in the function receiving the address. This
change handles this for cases where the address is passed to a direct call,
a broker function, or as a pass-through argument of a broker function.

Change-Id: I57375adf7ecfc0a797ec5fd9fd24c308be9ccb99
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308831
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>