[DTrans] CMPLRLLVM-25425: Update DTransFieldModRef query behavior to handle functions passed as parameters.
Because address taken functions are now being allowed as function parameters
for DTransFieldModRef analysis, the query function needs to be extended to
merge in the MOD/REF behavior of these calls when walking the list of functions
that may be called. Also, adjusted the output format of the test functionality
to improve the reliability of LIT tests being able to match the MOD/REF
results with specific queries.

Change-Id: I42f6247150ec8e82bd3be52b1b885f615291471c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308832
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>