CMPLRLLVM-25752: Disable unrolling of inner do-multi-exit loops in complicated functions.
The hottest function in 500.perlbench (S_regmatch) is absolutely huge.
Unrolling of do-multi-exit loops in this function is causing degradation
as reported in the JR. It is not known what exactly is the cause of
degradation. Unrolling could be causing more spills or it could be due
to alignment issues. The current patch is using some heuristics to
disable unrolling.

Change-Id: Ib9210943090086e0f61353e53e7d378052ed09b4
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308898
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>