[VPlan][IR CG] Handle addrspace mismatch while widening private allocas.
Privatization in LLVM-IR vector codegen is done by emitting a
widened alloca in default module addrspace. However this default
value need not match with addrspace that the private pointer is
declared with in OMP SIMD PRIVATE clause. This patch resolves the
mismatch by fixing CG to emit an explicit addrspacecast after
creating the wide alloca, if needed.

JIRA : CMPLRLLVM-25873

Change-Id: I111511cde545efd66b2c1d1cf83718fcf19dcd0c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/309159
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>