CMPLRLLVM-25935: clean-up VPO directives with optimizations enabled.
The current pipeline is quite messy. When optimizations are enabled
and Intel proprietary optimizations are disabled (e.g. SPIR-V offload
with pre-SPIR-V optimizations) we are not deleting VPO directives
that cause llvm-spirv crash. It looks like we need to straighten
the VPO passes pipeline. This is a conservative fix for merging
into update 1.

Change-Id: Ia36d36f39c6e7cb475f9fb91f3b9b10dbc0b9b69
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/309889
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>