[Driver][SYCL] Fix SYCL option passing for preprocessing (#3149)
When performing SYCL offload, various -fsycl -fsycl-is* options are passed
to the host and target compilations respectively.  When performing
preprocessing, a few of these options were not properly passed causing
problems with the generated preprocessed file for the host compilation
side.

addresses: https://jira.devtools.intel.com/browse/CMPLRLLVM-25244
This is a cherry-pick of: https://github.com/intel/llvm/pull/3149

Change-Id: Ica89c2594cd0c351f29546deab64a5a6eeedf18a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/310141
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Chilamantula, Harini <harini.chilamantula@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>