Mark the pass manager default as an Intel customization
This change is intended to force the CMake setting for enabling the new
pass manager by default to be flagged as a conflict when it is changed
by the LLVM community.

The current default is to use the legacy pass manager, and we would like
to keep using the legacy pass manager in xmain for the time being. With
this change, we should be prompted to make that decision during the
pulldown process.

Change-Id: I56667eae5ef0cb71bea64992311f19f12f9f30ed
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/310337
Reviewed-by: Vodopyanov, Dmitry <dmitry.vodopyanov@intel.com>
Tested-by: Vodopyanov, Dmitry <dmitry.vodopyanov@intel.com>