CMPLRLLVM-25955: Refine edges using path analysis.
When DDG is invoked on a region, some refs may be in mutually
exclusive paths. For example they can be inside the if-else cases
or inside different switch cases. We can skip creating edges in
this case. We also can skip creating backward edge if there is
no common loop for the refs inside the region.

The problem was encountered in the mentioned JR. Although the main
culprit was vectorizer invoking DDG on the region, the current
change helps to cut down the compile time to about 40% of the
original without fixing the vectorizer. The compile time is still
too big so the main fix is for vectorizer to not invoke DDG for the
region.

This fix will help in scaling DDG build for the region and may be
necessary when more loop fusion is triggered.

Please refer to JR comments to learn more about the compile time issue.

Change-Id: Ibb517c3f9dbb93cefbbe619202ac0dee77194029
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/310342
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: Chawla, Pankaj <pankaj.chawla@intel.com>