[LLD][COFF] CMPLRLLVM-25679: Include the IMAGE_REL_BASED_HIGHLOW relocation base type when the machine is 64 bits and the relocation type is ADDR32
The COFF driver produces an ABSOLUTE relocation base for an ADDR32
relocation type and the system is 64 bits (machine=AMD64). The
relocation information won't be added in the output and could
produce an incorrect address access during run-time. This change
set checks if the relocation type is IMAGE_REL_AMD64_ADDR32 and
if so, adds the relocated symbol as IMAGE_REL_BASED_HIGHLOW base.

NOTE: This changes are proposed to the community and are already
in discussion. Once the fix is in llorg then we can remove the
INTEL_CUSTOMIZATION tags.

Change-Id: Ic5a3300a1823ef3b6beb0e9ce04719fd1ebdf27d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/310519
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>