[Paropt] PRIVATE:NONPOD can have null ctor/dtor
The WRN builder must handle the case when the ctor/dtor function pointers
of a PRIVATE:NONPOD clause are "i8* null" in the IR. This can happen,
for example, during device compilation for directives outside of a TARGET
region. It is valid IR because we don't emit device code for such directives.

Change-Id: Iae1b97b57ee6582186d1d7bc228defd11821b885
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/310389
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>