[Paropt] Provide AA results to WRegionInfo with LegacyAARGetter()
ModulePass VPOParopt depends on FunctionPass WRegionInfoWrapperPass,
which transitively depends on TBAA, an ImmutablePass. Since ModulePasses
are not supposed to use FunctionPasses which depend on ImmutablePasses,
the end result seems to be that TBAA is not available in the resulting
AA pipeline.

However, using LegacyAARGetter() to fetch an AAResults from directly
within VPOParopt seems to yield a fully-populated AA pipeline. This
change takes advantage of this, populating the WRegionInfo's AAResults
via LegacyAARGetter() in VPOParopt's WRegionInfoGetter().

Change-Id: I1edb85dfb9839660c0a60bab434e5cfee11bc06b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/310585
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>