[1/2][InjectTLIMappings] Add test to reveal a shortcoming in addMappingsFromTLI.
Currently while populating vector-function variants from TLI,
declaration of a variant function is not made "sticky" if
the function is already declared in the module. This can lead
to scenarios where the declaration is removed later in pipeline
by LTO. This patch adds a test to reveal this shortcoming.

JIRA : CMPLRLLVM-24371

Change-Id: I4af0e6f2345b1b5258f9634db4d3b5f1a28ccc2c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/310825
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>