[SYCL] Temporarily restore event deps in in-order queues for Level Zero
Some of the enqueued operations are implemented directly in the Level
Zero plugin (e.g. map/unmap) and their implementation does not satisfy
in-order queue requirements, leading to incorrect behaviour if the event
dependencies that should be redundant are dropped.

Restore these event dependencies until this is fixed in the Level Zero
plugin.

Change-Id: I83978b42fb0e0b6bcbb7b224bd0d88009d63b181
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/310898
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>