[SYCL] Temporarily restore event deps in in-order queues for Level Zero
Some of the enqueued operations are implemented directly in the Level
Zero plugin (e.g. map/unmap) and their implementation does not satisfy
in-order queue requirements, leading to incorrect behaviour if the event
dependencies that should be redundant are dropped.

Restore these event dependencies until this is fixed in the Level Zero
plugin.

Change-Id: Ic51ff3f23f9897b4761479bbb42627230ac573d7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/310902
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>