CMPLRLLVM-25180: Extend inlining for MorphologyPrimitive
This extends the handling of dynamic allocas in the inlining heuristics to
tolerate the special dynamic alloca that is created for OMP parallel loops
in the "QUAL.OMP.JUMP.TO.END.IF" clause, and to tolerate dynamic allocas
which are character strings.

The code to recognize the special dynamic alloca that is created for OMP
parallel loops in the "QUAL.OMP.JUMP.TO.END.IF" clause was written by
Abhinav Gaba.

These changes are needed to get MorphologyPrimitive to inline in 638.imagick,
which then allows the cloning of MorphologyApply to be performed.

Change-Id: I568651343134b551281f7ba234853b42dafacc8d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/311038
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>