[LLD][ELF] CMPLRLLVM-26105: Fix for R_386_GOTOFF type in non-alloc relocation
The libraries from GCC 10x uses R_386_GOTOFF relocation type when
adding debug information. This produces an issue with the relocation
process since it doesn't understand the type.

This change set fixes the issue by allowing to process the new
relocation types for 32 bits.

This change is cherry picked from the community:
https://reviews.llvm.org/D95994

Change-Id: I5bed0a681a1e8bb08f61683ca845c59094d87cfb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/311255
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>