[BasicAA] Exclude aliasSelect logic from loopCarriedAlias
We accidentally allowed aliasSelect logic to contribute to
loopCarriedAlias with change 311397.

With this change, the check for NeedLoopCarried is moved to within
aliasSelect itself, which should make us insensitive to upstream changes
to where it is called.

This change also restores the original intent of the HIRDDAnalysis LIT
tests which checked that loopCarriedAlias does not incorrectly indicate
that we can break DDG edges involving select operations.

Change-Id: Iff47aae64c866538d3671bec6ee33e2117046310
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/312023
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>