[SYCL] Store command lists cache in the context class
Currently command lists cache is stored in the device class and command
lists are destroyed only during piTearDown during destruction of device
objects but this is too late because associated contexts of the command
lists aren't alive by this time.
It looks reasonable to store the cache in the context class and destroy
command lists associated with a context during piContextRelease because
command lists are not usable after context destruction.

Change-Id: Ic82748db331488f6c3765a68d5bffb569d949c7b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/311664
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>