[SYCL] Store command lists cache in the context class
Currently command lists cache is stored in the device class and command
lists are destroyed only during piTearDown during destruction of device
objects but this is too late because associated contexts of the command
lists aren't alive by this time.
It looks reasonable to store the cache in the context class and destroy
command lists associated with a context during piContextRelease because
command lists are not usable after context destruction.

Change-Id: I6674ef97b989333c1908395ac1f6383a1920cc2a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/311662
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Maslov, Sergey V <sergey.v.maslov@intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>