CMPLRLLVM-26177: Add missing handling for GlobalIFuncs
The module summary analysis and bitcode readers and writers did not
completely imnplement handling for GlobalIFuncs. In particular, the
fact that the Resolver of a GlobalIFunc is a referenced by the GlobalIFunc
was missed, causing improper dead function elimination when the Resolver is
not a preserved symbol during LTO.

I have also fixed two issues is Globals.cpp and LTOModule.cpp related to
GlobalIFuncs that showed up when I was writing the LIT test. Also, the
LIT test intel_lto-ifunc_01.ll has to be generalized.

Change-Id: I478d6e38499840105d97d85b0bff54ff054b73ad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/312370
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>