[NFC] CMPLRLLVM-26232: Separate the devirt multiversioning to an Intel proprietary file (part 1)
This change set is the first part in an effort to have a better
modularity for the multiversioning in the whole program
devirtualization. We need to reduce the data dependency in the
multiversioning from the WP-devirt structures in order to move
proprietary changes into another file. This change set introduces
a new structure that holds the basic information needed to apply
the multiversioning. A new function will collect the virtual call
sites and the targets from the special structures, and group them
into the new structure that will be used to generate the multiversioning.

Change-Id: I01f212700539a680d3f12fedd32a4e75d82900b3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/312498
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>