[X86][AMX] Lower tile copy instruction.
Since there is no tile copy instruction, we need to store tile
register to stack and load from stack to another tile register.
We need extra GR to hold the stride, and we need stack slot to
hold the tile data register. We would run this pass after copy
propagation, so that we don't miss copy optimization. And we
would run this pass before prolog/epilog insertion, so that we
can allocate stack slot.

Differential Revision: https://reviews.llvm.org/D97112

Change-Id: I0bee329b48b76cd7b4af86dc6f05f1052c5e650d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/317518
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>