[VPlan][NFC] Avoid unnecessary use of parent's DDG in HIR VLS analysis
HIR VLS analysis was using the DDG of the parent of the loop being vectorized.
This can be compile time expensive as re-computing the DDG can be expensive
in cases where the analyses are marked as invalidated. This actually caused
a compilation to time-out in one of spec2000 tests where we were unnecessarily
computing the DDG of an outer loop several times as we vectorized several
inner loops causing the outer loop analyses to be invalidated.

Jira :: CMPLRLLVM-25955

Change-Id: I785904860d13aea529b587a11085cb7bde9b7828
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/312625
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>