Update versioning output to improve representation
The output of the versioning information (the more detailed section)
consisted of YYYY.m.PATCH.MMDD, which was better suited for the
frequent montly drops we had during the Beta.

Now that we are in Gold, these values need to be better presented
to reduce confusion.  New:  MAJOR.MINOR.PATCH.YYYYMMDD
Addresses:  https://jira.devtools.intel.com/browse/CMPLRLLVM-26400

Change-Id: I4a01c2154b2928d0efc976953366ef61b3bcc607
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/312904
Reviewed-by: Ammarguellat, Zahira <zahira.ammarguellat@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>