[VPlan][CallVecDecisions][NFC] Provide mechanism to reset CallVecDecisions results.
Results of CallVecDecisions analysis are persisent in the way that
they are recorded in corresponding VPCallInstruction rather than
a side temporary data structure. This leads to inconsistent state
of results in VPlan when the last analyzed VF does not correspond
to the final VF that is chosen by CM. In this patch we enable a
mechanism to reset CallVecDecisions results. Since this analysis
is tightly coupled with SVA, resetting is done whenever SVA results
are cleared/invalidated for a given VPlan. In a follow-up patch
CallVecDecisions invocation will also be moved to runSVA() for
consistency.

Change-Id: Id1339c5965153127a1e05de281a1eab5c6bdc86a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/312949
Reviewed-by: Zverev, Sergey I <sergey.i.zverev@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>