[Paropt] Don't try to emit critical section around UDR finish code for SIMD loops.
Reduction update code for scalar reductions had a check for SIMD loops
to prevent generation of kmpc_begin/end_critical/kmpc_reduce/kmpc_end_reduce
calls around the reduction finalization code.

This check was not propagated to the UDR code-gen when the support for UDR
was added, which was causing a compiler assertion. This commit fixes that.

Change-Id: I01e904a08aa65f82cee8c8fd99f9d023d13cce81
JIRA: CMPLRLLVM-26324, CMPLRLLVM-26470
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/313133
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>