[CMPLRLLVM-26637] ArgNoAliasProp: obtain analyses for callsites' functions
When considering all callsites for the candidate function we obtain and
use only the candidate function's DT/AA analyses even though we are
potentially analyzing various other functions containing callsites. This
has generally worked due to luck, but can cause incorrect results.

For example, if we pass DT analysis on the callee to CaptureTracking to
reason about reachability of a instruction in the caller we can
incorrectly conclude that instruction is not reachable. This can cause
us to miss capturing instructions.

With this change we query the needed analyses for each caller function
as we visit each callsite.

Change-Id: I34178a043b658c3adb3b3186033d0b63087d5b22
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/313277
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>