Revised changes for InliningLoopInfoCache in xmain-cand
I did not see the failures that Dmitry Vodopyanov saw, but the code in
xmain-cand does have some problems, so here are my suggested fixes.
All code dealing with the InliningLoopInfoCache should be under
INTEL_CUSTOMIZATION, since the code is only used in xmain, not llorg.
Also, I moved all code dealing with it to llvm/lib/Analysis.

Change-Id: I3fde33e38fb83569ff7a212d88469ceda30f7536
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/313368
Reviewed-by: Vodopyanov, Dmitry <dmitry.vodopyanov@intel.com>
Tested-by: Vodopyanov, Dmitry <dmitry.vodopyanov@intel.com>