[xmain] Enable transformation of sinf/cosf -> sinpif/cospif by default, when certain conditions hold. Also enable transformation of sin/cos -> sinpi/cospi under a compiler option.
Details:
This change implements updates to the transformation of sin to sinpi and
cos to cospi. With this change, the transformation is enabled by default
for the float versions of the math functions (sinf and cosf), when certain
conditions hold. These conditions are checked by the function
doConvertToSinpiOrCospi() in Intel_TransformSinAndCosCalls.cpp.
SPEC ACCEL benchmark, 514.pomriq, gets a performance boost with this
transformation.

The transformation is *not* enabled by default for the double versions
of the math functions (sin and cos), as more testing of accuracy of
results is needed. The compiler option '-mllvm -enable-transform-sin-cos-double'
can be used to enable the transformation for the double versions of the
math functions. To disable the transformation altogether, for both float
and double, use the compiler option '-mllvm -enable-transform-sin-cos=0'.

See:
https://jira.devtools.intel.com/browse/CMPLRLLVM-20297
https://git-amr-2.devtools.intel.com/gerrit/271031

Change-Id: I249d61886fb760123e2a4a536e300479a87236e5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/313372
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Copty, Nawal <nawal.copty@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>